#6167 – Fix cyclic dependencies for unit tests run
