+++ packages/ketcher-core/src/application/editor/tools/Bond.ts
@@ -21,9 +21,9 @@ import { SnakeModePolymerBondRenderer } from 'application/render/renderers/Polym
 import assert from 'assert';
 import { AttachmentPoint } from 'domain/AttachmentPoint';
 import {
+  AmbiguousMonomer,
   UnresolvedMonomer,
   UnsplitNucleotide,
-  AmbiguousMonomer,
 } from 'domain/entities';
 import { BaseMonomer } from 'domain/entities/BaseMonomer';
 import { Chem } from 'domain/entities/Chem';
@@ -39,17 +39,12 @@ import { AttachmentPointName } from 'domain/types';
 //  because of using uncontrolled `index.ts` files.
 import { Coordinates } from '../shared/coordinates';
 import { AtomRenderer } from 'application/render/renderers/AtomRenderer';
-import { ToolName } from 'application/editor';
+import { MACROMOLECULES_BOND_TYPES, ToolName } from 'application/editor';
 import { KetMonomerClass } from 'application/formatters';
 
 type FlexModeOrSnakeModePolymerBondRenderer =
   | FlexModePolymerBondRenderer
   | SnakeModePolymerBondRenderer;
-
-export enum MACROMOLECULES_BOND_TYPES {
-  SINGLE = 'single',
-  HYDROGEN = 'hydrogen',
-}
 
 class PolymerBond implements BaseTool {
   private bondRenderer?: FlexModeOrSnakeModePolymerBondRenderer;
