+++ packages/ketcher-core/src/application/render/renderers/PolymerBondRenderer/PolymerBondRendererFactory.ts
@@ -1,4 +1,4 @@
-import { CoreEditor, SnakeMode } from 'application/editor';
+import { CoreEditor } from 'application/editor';
 import { FlexModePolymerBondRenderer } from 'application/render/renderers/PolymerBondRenderer/FlexModePolymerBondRenderer';
 import { SnakeModePolymerBondRenderer } from 'application/render/renderers/PolymerBondRenderer/SnakeModePolymerBondRenderer';
 import { PolymerBond } from 'domain/entities/PolymerBond';
@@ -52,5 +52,5 @@ export class PolymerBondRendererFactory {
 
 function checkIfIsSnakeMode(): boolean {
   const editor = CoreEditor.provideEditorInstance();
-  return editor?.mode instanceof SnakeMode;
+  return editor?.mode.modeName === 'flex-layout-mode';
 }
